chore: move helper classes into separate files
